pyproject.toml
[tool.black]
line-length = 88
target_version = ['py37', 'py38', 'py39', 'py310', "py311"]
exclude = '''
(
  /(
    | \.git
    | \.venv
    | \.mypy_cache
    | \.tox
  )/
)
'''

[tool.isort]
line_length = 88
profile = "black"

[tool.poetry]
name = "ocp-addons-operators-cli"
version = "1.0.0"
description = "CLI to install/uninstall Addons/operators on Openshift clusters."
authors = ["Meni Yakove <myakove@gmail.com>", "Ruth Netser <ruth.netser@gmail.com>"]
packages = [{include = "ocp_addons_operators_cli"}]
include = [{path = "manifests/*"}]
readme = "README.md"
repository = "https://github.com/RedHatQE/ocp-addons-operators-cli"

[tool.poetry.scripts]
cli = "app.cli:main"

[tool.poetry.dependencies]
python = "^3.8"
openshift-cluster-management-python-wrapper = "^1.0.21"
colorlog = "^6.7.0"
openshift-python-wrapper = "^4.14.1"
openshift-python-utilities = "^5.0.0"
click = "^8.1.4"
pyaml-env = "^1.2.1"
python-simple-logger = "^1.0.7"

[tool.poetry.dev-dependencies]
ipython = "*"
ipdb = "*"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
